Fixes #29 #85
Problem:

bigip_user lib needed an overhaul to be usable, this has led to quite a few issues

Analysis:

Reworked bigip_user lib with the corresponding tests and docs, to be in line with the new convention. Added few safety checks to ensure that existing data does not get overwritten.

Tests:

Unit
Functional
Flake 8